+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/settings/RoomSettingsTopicViewModelTest.kt
@@ -176,6 +176,7 @@ class RoomSettingsTopicViewModelTest {
                 userId = UserId("test", "server"),
             ),
             selectedRoomId = roomId,
+            onOpenMention = { _, _  -> },
         )
     }
 }